+++ src/test/globalSetup.test.ts
@@ -17,7 +17,7 @@ import { getLogger, LogLevel } from '../shared/logger'
 import { setLogger } from '../shared/logger/logger'
 import { CodelensRootRegistry } from '../shared/sam/codelensRootRegistry'
 import { activateExtension } from '../shared/utilities/vsCodeUtils'
-import { FakeExtensionContext } from './fakeExtensionContext'
+import { FakeExtensionContext, FakeMemento } from './fakeExtensionContext'
 import { TestLogger } from './testLogger'
 import * as testUtil from './testUtil'
 
@@ -58,6 +58,7 @@ beforeEach(async function () {
     globals.telemetry.telemetryEnabled = true
     globals.telemetry.clearRecords()
     globals.telemetry.logger.clear()
+    ;(globals.context as FakeExtensionContext).globalState = new FakeMemento()
 
     await testUtil.closeAllEditors()
 })